Fix. rename useGetAllDocumens to singleDatabase based on its actual purpose and the data returned
