+++ src/apis/useGetSingleDatabase.jsx
@@ -2,37 +2,88 @@
 import { useQuery } from "@tanstack/react-query";
 import { useAtomValue, useSetAtom } from "jotai";
 
-import { currentDBIdAtom, userAtom, fieldsAtom } from "../atoms/atoms";
-
 import fetchData from "../utils/axios";
+
+import {
+  currentDBIdAtom,
+  userAtom,
+  documentsNumAtom,
+  documentsIdsAtom,
+  changedDocAtom,
+  docDataAtom,
+  primaryFieldAtom,
+  relationshipsDataAtom,
+  fieldsAtom,
+} from "../atoms/atoms";
 
 import Loading from "../components/shared/Loading";
 
 function useGetSingleDatabase() {
   const { userId } = useAtomValue(userAtom);
   const currentDBId = useAtomValue(currentDBIdAtom);
+
+  const setDocumentsNum = useSetAtom(documentsNumAtom);
+  const setDocumentsIds = useSetAtom(documentsIdsAtom);
+  const setChangedDoc = useSetAtom(changedDocAtom);
+  const setDocData = useSetAtom(docDataAtom);
+  const setPrimaryField = useSetAtom(primaryFieldAtom);
+  const setRelationshipsData = useSetAtom(relationshipsDataAtom);
   const setFields = useSetAtom(fieldsAtom);
 
-  async function getDatabase() {
+  async function getSingleDatabase() {
     const response = await fetchData(
       "GET",
       `users/${userId}/databases/${currentDBId}`,
     );
 
-    return response.data.database.documents[0];
+    return response.data.database;
   }
 
-  const { isLoading } = useQuery(["userDb", currentDBId], getDatabase, {
-    enabled: !!userId && !!currentDBId,
-    onSuccess: result => {
-      setFields(result.fields);
+  const { data: singleDatabase, isLoading } = useQuery(
+    ["SingleDatabase", currentDBId],
+    getSingleDatabase,
+    {
+      retry: false,
+      enabled: !!userId && !!currentDBId,
+      onSuccess: result => {
+        const documentsId = [];
+
+        const docs = result.documents.map(document => {
+          documentsId.push(document._id);
+
+          return { documentId: document._id, fields: [] };
+        });
+
+        setDocData(result.documents);
+        setFields(result.documents[0].fields);
+        setChangedDoc(docs);
+        setDocumentsIds(documentsId);
+        setDocumentsNum(result.documents.length);
+
+        if (result.relationships?.length) {
+          setRelationshipsData(result.relationships);
+
+          const primaryFieldsList = result.relationships.map(element => {
+            return element.primaryFieldName;
+          });
+
+          setPrimaryField(primaryFieldsList);
+
+          return;
+        }
+
+        setRelationshipsData(null);
+      },
+      refetchOnWindowFocus: false,
+      staleTime: Infinity,
     },
-    refetchOnWindowFocus: false,
-  });
+  );
 
   if (isLoading) {
     return <Loading />;
   }
+
+  return { singleDatabase };
 }
 
 export default useGetSingleDatabase;
