+++ src/components/Modals/Relationship/Done.jsx
@@ -31,7 +31,7 @@ function Done() {
       <Button
         className="w-20 h-8 mt-5 rounded-md bg-black-bg text-white hover:bg-dark-grey"
         onClick={() => {
-          queryClient.refetchQueries(["DocumentsList", currentDBId]);
+          queryClient.refetchQueries(["SingleDatase", currentDBId]);
           queryClient.refetchQueries(["Relationships", currentDBId]);
           setShowRelationshipModal(false);
           setRelationshipStep("start");
